Re add supertab to vim
Originally had removed supertab to find a bug which happened when I had
an autocomplete suggestion when standing at the end of a line, and I
wanted to press enter to move onto the next line. The problem was that
when I pressed enter I inserted the suggestion instead of making a new
line. I thought this was the blame of supertab so I removed it.

Turns out its vim's blame, and I am not sure it can be fixed. So I am re
adding supertab.